Moving folder and follow guideline
