+++ domain/src/main/java/com/dana/domain/base/UseCaseSchedulers.kt
@@ -1,4 +1,4 @@
-package com.dana.domain.usecase
+package com.dana.domain.base
 
 import io.reactivex.Scheduler
 import io.reactivex.android.schedulers.AndroidSchedulers
