+++ domain/src/main/java/com/dana/domain/listrecipe/model/Recipe.kt
@@ -1,0 +1,7 @@
+package com.dana.domain.listrecipe.model
+
+/**
+ * @author Iga Noviyanti created on 22/04/22 at 09.51
+ */
+class Recipe {
+}
