+++ domain/src/main/java/com/dana/domain/listrecipe/repository/RecipeRepository.kt
@@ -1,9 +1,7 @@
-package com.dana.domain.repository
+package com.dana.domain.listrecipe.repository
 
 import io.reactivex.Observable
 
 interface RecipeRepository {
     fun getListRecipe(from: Int, size: Int) : Observable<Int>
 }
-
-//interface GetListRecipeCacheRepository
