--- domain/src/main/java/com/dana/domain/model/BaseModel.kt
@@ -1,9 +1,0 @@
-package com.dana.domain.model
-
-/**
- * @author Iga Noviyanti created on 18/04/22 at 13.13
- */
-data class BaseModel (
-    val name: String? = null,
-    val size: Int? = null
-)
